Minor corrections

+++ Wasserstein/RunAll.py
@@ -40,8 +40,8 @@ use_hedging = False
 use_batchnorm = False
 
 # simulation parameters
-batch_size = 40000
-epochs = 100000
+batch_size = 20000
+epochs = 1000
 N_simulations = 20 * batch_size
 N_steps = 96
 period_length = N_steps // n_maturities
@@ -66,44 +66,47 @@ if not os.path.exists('Data/target_Wasserstein.pth.tar'):
 else:
     target = torch.load('Data/target_Wasserstein.pth.tar')
 
-with open("Results/log_eval_LV_Wasserstein.txt", "w") as f:
-    f.write('epoch,loss\n')
+if not os.path.exists('Results/Wasserstein_NeuralLV.pth.tar'):
+    with open("Results/log_eval_Wasserstein_NeuralLV.txt", "w") as f:
+        f.write('epoch,loss\n')
 
-modelLV = NeuralLV(device=device, batch_size=batch_size, dropout=dropout, use_batchnorm=use_batchnorm, use_hedging=use_hedging,
-                   N_simulations=N_simulations, N_steps=N_steps, Time_horizon=Time_horizon, period_length=period_length,
-                   S0=S0, n_maturities=n_maturities, n_strikes=n_strikes, rfr=rfr,
-                   num_layers=num_layers, layer_size=layer_size,
-                   num_layers_hedging=num_layers_hedging, layer_size_hedging=layer_size_hedging,
-                   test_normal_variables=test_normal_variables)
+    modelLV = NeuralLV(device=device, batch_size=batch_size, dropout=dropout, use_batchnorm=use_batchnorm, use_hedging=use_hedging,
+                       N_simulations=N_simulations, N_steps=N_steps, Time_horizon=Time_horizon, period_length=period_length,
+                       S0=S0, n_maturities=n_maturities, n_strikes=n_strikes, rfr=rfr,
+                       num_layers=num_layers, layer_size=layer_size,
+                       num_layers_hedging=num_layers_hedging, layer_size_hedging=layer_size_hedging,
+                       test_normal_variables=test_normal_variables)
 
-print('Neural Local Volatility Model initiated')
-train_Wasserstein(modelLV, target, epochs, batch_size)
-print('Neural Local Volatility Model trained')
-torch.cuda.empty_cache()
+    print('Neural Local Volatility Model initiated')
+    train_Wasserstein(modelLV, target, epochs, batch_size)
+    print('Neural Local Volatility Model trained')
+    torch.cuda.empty_cache()
 
-with open("Results/log_eval_Wasserstein_NeuralLSV.txt", "w") as f:
-    f.write('epoch,loss\n')
+if not os.path.exists('Results/Wasserstein_NeuralLSV.pth.tar'):
+    with open("Results/log_eval_Wasserstein_NeuralLSV.txt", "w") as f:
+        f.write('epoch,loss\n')
 
-modelLSV = NeuralLSV(device=device, batch_size=batch_size, dropout=dropout, use_batchnorm=use_batchnorm, use_hedging=use_hedging,
-                     N_simulations=N_simulations, N_steps=N_steps, Time_horizon=Time_horizon, period_length=period_length,
-                     S0=S0, n_maturities=n_maturities, n_strikes=n_strikes, rfr=rfr,
-                     num_layers=num_layers, layer_size=layer_size,
-                     num_layers_hedging=num_layers_hedging, layer_size_hedging=layer_size_hedging,
-                     test_normal_variables=test_normal_variables)
-print('Neural Local Stochastic Volatility Model initiated')
-train_Wasserstein(modelLSV, target=target, batch_size=batch_size, epochs=epochs, threshold=2e-5)
-print('Neural Local Stochastic Volatility Model trained')
-torch.cuda.empty_cache()
+    modelLSV = NeuralLSV(device=device, batch_size=batch_size, dropout=dropout, use_batchnorm=use_batchnorm, use_hedging=use_hedging,
+                         N_simulations=N_simulations, N_steps=N_steps, Time_horizon=Time_horizon, period_length=period_length,
+                         S0=S0, n_maturities=n_maturities, n_strikes=n_strikes, rfr=rfr,
+                         num_layers=num_layers, layer_size=layer_size,
+                         num_layers_hedging=num_layers_hedging, layer_size_hedging=layer_size_hedging,
+                         test_normal_variables=test_normal_variables)
+    print('Neural Local Stochastic Volatility Model initiated')
+    train_Wasserstein(modelLSV, target=target, batch_size=batch_size, epochs=epochs, threshold=2e-5)
+    print('Neural Local Stochastic Volatility Model trained')
+    torch.cuda.empty_cache()
 
-with open("Results/log_eval_Wasserstein_NeuralSDE.txt", "w") as f:
-    f.write('epoch,loss\n')
+if not os.path.exists('Results/Wasserstein_NeuralSDE.pth.tar'):
+    with open("Results/log_eval_Wasserstein_NeuralSDE.txt", "w") as f:
+        f.write('epoch,loss\n')
 
-modelSDE = NeuralSDE(device=device, batch_size=batch_size, dropout=dropout, use_batchnorm=use_batchnorm, use_hedging=use_hedging,
-                     N_simulations=N_simulations, N_steps=N_steps, Time_horizon=Time_horizon, period_length=period_length,
-                     S0=S0, n_maturities=n_maturities, n_strikes=n_strikes, rfr=rfr,
-                     num_layers=num_layers, layer_size=layer_size,
-                     num_layers_hedging=num_layers_hedging, layer_size_hedging=layer_size_hedging,
-                     test_normal_variables=test_normal_variables)
-print('Neural Stochastic Differential Equation Model initiated')
-train_Wasserstein(modelSDE, target=target, batch_size=batch_size, epochs=epochs, threshold=2e-5)
-print('Neural Stochastic Differential Equation Model trained')
+    modelSDE = NeuralSDE(device=device, batch_size=batch_size, dropout=dropout, use_batchnorm=use_batchnorm, use_hedging=use_hedging,
+                         N_simulations=N_simulations, N_steps=N_steps, Time_horizon=Time_horizon, period_length=period_length,
+                         S0=S0, n_maturities=n_maturities, n_strikes=n_strikes, rfr=rfr,
+                         num_layers=num_layers, layer_size=layer_size,
+                         num_layers_hedging=num_layers_hedging, layer_size_hedging=layer_size_hedging,
+                         test_normal_variables=test_normal_variables)
+    print('Neural Stochastic Differential Equation Model initiated')
+    train_Wasserstein(modelSDE, target=target, batch_size=batch_size, epochs=epochs, threshold=2e-5)
+    print('Neural Stochastic Differential Equation Model trained')
